Make Fragment.requestable? depend only on respond_to?(:request).
This involves removing Fragment.request so that a subclass will only
respond to :request if it defines it itself. The class method also no
longer requires respond_to?(:request_path) and no longer incorporates the
instance method requestable?, which requires an additional change to
the expression for the condition under which Fragment.request_queues
returns nil.

Also remove the root class test from Fragment.root - in general a root
fragment need not be requestable. There is no fundamental way to
distinguish a subclass that may be used as a root fragment.