Move parsing of ApplicationInstance into the UserSession
There is an incompatibility between the parser used in URI.parse in
Ruby 2.1.7 and 2.4.9. This is a problem when the request queue is created
in one application instance and processed in another, each with different
Rubies.

The UserSession is instantiated in the application instance in which the
queue is to be processed, so we can parse the url there without a problem.