Use custom callback for app request after creating a record
When a class declares 'needs_record_id' or 'set_record_type' (for some
subclasses whose parent already has 'needs_record_id', if the class is
'requestable' i.e. it defines a 'request_path' that allows the fragment
content to be generated by sending a 'get' request to the app, when a new
record is created we automaticlly send the request defined for the
fragment class to the app.

Previously this was done by calling 'subscribe_to' on the class and
defining a create_#{record_class.model_name.param_key}_successful method
in which the request is created and placed in the request queue.

However this approach is problematic if the fragment class in question is
a subclass of another one that defines a method of the same name within an
'inherited' callback. The method created as a result of 'set_record_type' will replace the
version created by the 'inherited' callback.

This patch solves this problem by creating the app request in a custom
callback attached to the 'after_create' method on the record class
subscription for the record class to instead of 'subscribe_to'.