Ensure that Fragment#touch_or_destroy does not generate app requests
The :no_request option on #touch was not being propagated to #touch_parent
via the after_commit callback. We add an instance variable @no_request to
store the setting and reset it once the parent has been touched.

This option is now also available on #destroy, since this also triggers a
call to #touch_parent, and this is used in the invocation of #destroy in
#touch_or_destroy.