Add the current queue_suffix to the params sent with application requests.
This is to support a case where a request sent asynchronously spawns
a secondary request. It's important that the second request be sent to the
same ActiveJob backend queue as the one that spawned it to avoid the
test user's credentials being modified when signing in for the secondary
request while the process that queued the first request is still running.

Also in the use case that made  this necesary, bulk updating of content
after a view change, it keeps the secondary requests off the default queue
that is used for higher priority requests after a user creates/updates
content.